Teach test_decoding plugin to work with 2PC
Implement all callbacks required for decoding 2PC in this test_decoding
plugin. Includes relevant test cases as well.

Additionally, includes a new option "check-xid". If this option points
to a valid xid, then the pg_decode_change() API will wait for it to
be aborted externally. This allows us to test concurrent rollback of
a prepared transaction while it's being actually decoded simultaneously.