Change tablespace of partitioned indexes during REINDEX.
There are some doubts about proper locking of partitions
here. AccessExclusiveLock is surely enough, but may be
a reason of deadlock. While ShareUpdateExclusiveLock
or ShareLock may be not sufficient.

Citing Michael:

"Are you sure that this does not represent a risk of deadlocks as EAL
is not taken consistently across all the partitions?  A second issue
here is that this breaks the assumption of REINDEX CONCURRENTLY kicked
on partitioned relations that should use ShareUpdateExclusiveLock for
all its steps.  This would make the first transaction invasive for the
user, but we don't want that."